Removed repos from orm/Moved logic from database file to orm file
