+++ services/assistant/app/database/__init__.py
@@ -1,4 +1,3 @@
-from app.database.orm import AbstractRepository
-from app.database.database import get_db_factory
+from app.database.orm import get_db_factory
 
-__all__ = ["AbstractRepository", "get_db_factory"]
+__all__ = ["get_db_factory"]
